feat : add global exception handling and custom exceptions framework

+++ src/main/kotlin/team/twelve/ahu/domain/feed/service/DeleteFeedService.kt
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import org.springframework.web.bind.annotation.PathVariable
 import team.twelve.ahu.domain.feed.entitiy.repository.FeedRepository
+import team.twelve.ahu.global.exception.EntityNotFoundException
 import java.util.UUID
 
 @Service
@@ -12,6 +13,9 @@ class DeleteFeedService(
 ) {
     @Transactional
     fun execute(@PathVariable id: UUID) {
+        if (!feedRepository.existsById(id)) {
+            throw EntityNotFoundException("ID가 ${id}인 피드를 찾을 수 없습니다.")
+        }
         feedRepository.deleteById(id)
     }
 }
